requirements.txt
streamlit>=1.35.0
plotly>=5.18.0
rdkit>=2023.9.1
pandas>=2.0.0
Pillow>=10.0.0
